fix imports module matModuleMaterial
