+++ src/app/app.component.html
@@ -1,20 +1,1 @@
-<div>
-  <input 
-    #inputFile 
-    type="file" 
-    multiple
-    hidden="true"
-    accept=".jpg, .jpeg, .png"
-    (change)="fileInputchange($event)"
-  />
-  <button 
-    mat-fab color="primary"
-    (click)="inputFile.click()"
-    >+
-  </button>
-
-  <div #images> </div>
-
-  <button mat-raised-button>Upload</button>
-</div>
 <shared-image-previewer></shared-image-previewer>
