+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component, ElementRef, Renderer2, ViewChild } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
@@ -6,29 +6,5 @@ import { Component, ElementRef, Renderer2, ViewChild } from '@angular/core';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  @ViewChild('images') images: ElementRef;
-
-  constructor( private renderer: Renderer2 ) { }
-
-  fileInputchange(event) {
-
-    for (let fileKey in event.target.files) {
-      const reader: FileReader = new FileReader();
-
-      if (!isNaN(parseInt(fileKey))) {
-        reader.readAsDataURL(event.target.files[fileKey]);
-
-        reader.onload = () => {
-          const image: ElementRef = this.renderer.createElement('img');
-        
-          this.renderer.setProperty(image, 'src', reader.result);
-          this.renderer.appendChild(this.images.nativeElement, image);
-
-          this.renderer.setStyle(image, 'width', '150px');
-          this.renderer.setStyle(image, 'height', '200px');
-        }  
-      }
-    };
-  }
-
+  
 }
