+++ src/app/app.module.ts
@@ -1,6 +1,5 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
 
 import { AppComponent } from './app.component';
 import { SharedModule } from '../shared/shared.module';
@@ -11,7 +10,6 @@ import { SharedModule } from '../shared/shared.module';
   ],
   imports: [
     BrowserModule,
-    MatButtonModule,
     SharedModule
   ],
   providers: [],
