+++ src/shared/components/image-previewer/image-previewer.component.ts
@@ -1,15 +1,36 @@
-import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+import { Component, ViewChild, ElementRef, Renderer2 } from '@angular/core';
 
 @Component({
   selector: 'shared-image-previewer',
   templateUrl: './image-previewer.component.html',
   styleUrls: ['./image-previewer.component.scss']
 })
-export class ImagePreviewerComponent implements OnInit {
+export class ImagePreviewerComponent {
 
-  constructor() { }
+  @ViewChild('images') images: ElementRef;
 
-  ngOnInit(): void {
+  constructor( private renderer: Renderer2 ) { }
+
+  fileInputchange(event) {
+
+    for (let fileKey in event.target.files) {
+      const reader: FileReader = new FileReader();
+
+      if (!isNaN(parseInt(fileKey))) {
+        reader.readAsDataURL(event.target.files[fileKey]);
+
+        reader.onload = () => {
+          const image: ElementRef = this.renderer.createElement('img');
+        
+          this.renderer.setProperty(image, 'src', reader.result);
+          this.renderer.appendChild(this.images.nativeElement, image);
+
+          this.renderer.setStyle(image, 'width', '150px');
+          this.renderer.setStyle(image, 'height', '200px');
+        }  
+      }
+    };
   }
+
 
 }
